==> scripts/01-download_data.R <==
#### Preamble ####
# Purpose: Downloads and saves the data from https://data.cdc.gov
# Author: Sarah Lee
# Date: 8 April 2024
# Contact: sarahhhh.lee@mail.utoronto.ca 
# License: MIT


### Download Data ###
# Proceed to your web browser and paste this URL "https://data.cdc.gov" 
# Search in the CDC website "Death rates for suicide, by sex, race, Hispanic origin, and age: United State"
# Click on the dataset with the name "Death rates for suicide, by sex, race, Hispanic origin, and age: United States". The dataset selected should be provided by NCHS.
# Click on the "Export" button located on the top right of the page
# Select the "Download file", then make sure to export data in CSV form and select "All data" if not selected
# Click Download

### Downloaded Data ###
# Save downloaded CSV into "data/raw_data/Death_rates_for_suicide__by_sex__race__Hispanic_origin__and_age__United_States_20240410.csv"
